+++ lightsim2grid/gridmodel/_aux_add_slack.py
@@ -1,15 +1,17 @@
-# Copyright (c) 2020, RTE (https://www.rte-france.com)
+# Copyright (c) 2020-2023, RTE (https://www.rte-france.com)
 # See AUTHORS.txt
 # This Source Code Form is subject to the terms of the Mozilla Public License, version 2.0.
 # If a copy of the Mozilla Public License, version 2.0 was not distributed with this file,
 # you can obtain one at http://mozilla.org/MPL/2.0/.
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of LightSim2grid, LightSim2grid implements a c++ backend targeting the Grid2Op platform.
+
 import warnings
 import numpy as np
+from ._pp_bus_to_ls_bus import pp_bus_to_ls
 
 
-def _aux_add_slack(model, pp_net):
+def _aux_add_slack(model, pp_net, pp_to_ls):
     """
     add the slack bus(es) to the lightsim2grid "model" based on the information in the pandapower network.
 
@@ -49,7 +51,7 @@ def _aux_add_slack(model, pp_net):
             
         # in this case the ext grid is not taken into account, i raise a warning if
         # there is one
-        slack_bus_ids = pp_net.ext_grid["bus"].values
+        slack_bus_ids = pp_bus_to_ls(pp_net.ext_grid["bus"].values, pp_to_ls)
         if pp_net.ext_grid.shape[0] >= 1:
             warnings.warn("LightSim will not consider the pandapower \"ext_grid\" as there "
                           "are already generators tagged as slack bus")
@@ -59,15 +61,15 @@ def _aux_add_slack(model, pp_net):
                       "I will attempt to add some from the ext_grid.")
         # first i try to see if a generator is connected to a slack bus
         # TODO SLACK: deactivate warnings (will be done at the end when more tests will be done)
-        slack_bus_ids = pp_net.ext_grid["bus"].values
+        slack_bus_ids = pp_bus_to_ls(pp_net.ext_grid["bus"].values, pp_to_ls)
         # if pp_net.ext_grid.shape[0] >= 2:
         #     warnings.warn("LightSim cannot handle multiple slack bus at the moment. Only the first "
         #                   "slack bus of pandapower will be used.")
 
-        if np.all(np.isin(slack_bus_ids, pp_net.gen["bus"].values)):
+        if np.all(np.isin(slack_bus_ids, pp_bus_to_ls(pp_net.gen["bus"].values, pp_to_ls))):
             # all slack buses have a generator connected to them
             # so i assume it was just a computation artifact, and assign these generators as slack buses
-            slack_gen_ids = np.isin(pp_net.gen["bus"].values, slack_bus_ids)  # id of generators connected to slack bus
+            slack_gen_ids = np.isin(pp_bus_to_ls(pp_net.gen["bus"].values, pp_to_ls), slack_bus_ids)  # id of generators connected to slack bus
             slack_gen_ids = np.where(slack_gen_ids)[0]  # keep only the id of the generators
             if "slack_weight" in pp_net.gen:
                 slack_coeff = pp_net.gen["slack_weight"].values[slack_gen_ids]
@@ -100,7 +102,7 @@ def _aux_add_slack(model, pp_net):
             vm_pu = 1.0 * pp_net.ext_grid["vm_pu"].values
             gen_p = np.concatenate((pp_net.gen["p_mw"].values, slack_contrib))
             gen_v = np.concatenate((pp_net.gen["vm_pu"].values, vm_pu))
-            gen_bus = np.concatenate((pp_net.gen["bus"].values, slack_bus_ids))
+            gen_bus = np.concatenate((pp_bus_to_ls(pp_net.gen["bus"].values, pp_to_ls), slack_bus_ids))
             gen_min_q = np.concatenate((pp_net.gen["min_q_mvar"].values, [-999999.]))
             gen_max_q = np.concatenate((pp_net.gen["max_q_mvar"].values, [+99999.]))
             model.init_generators(gen_p, gen_v, gen_min_q, gen_max_q, gen_bus)