+++ lightsim2grid/gridmodel/_aux_add_trafo.py
@@ -1,15 +1,17 @@
-# Copyright (c) 2020, RTE (https://www.rte-france.com)
+# Copyright (c) 2020-2023, RTE (https://www.rte-france.com)
 # See AUTHORS.txt
 # This Source Code Form is subject to the terms of the Mozilla Public License, version 2.0.
 # If a copy of the Mozilla Public License, version 2.0 was not distributed with this file,
 # you can obtain one at http://mozilla.org/MPL/2.0/.
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of LightSim2grid, LightSim2grid implements a c++ backend targeting the Grid2Op platform.
+
 import warnings
 import numpy as np
+from ._pp_bus_to_ls_bus import pp_bus_to_ls
 
 
-def _aux_add_trafo(converter, model, pp_net):
+def _aux_add_trafo(converter, model, pp_net, pp_to_ls):
     """
     Add the transformers of the pp_net into the lightsim2grid "model"
 
@@ -89,8 +91,8 @@ def _aux_add_trafo(converter, model, pp_net):
                      tap_pos,
                      shift_,
                      is_tap_hv_side,
-                     pp_net.trafo["hv_bus"].values,
-                     pp_net.trafo["lv_bus"].values)
+                     pp_bus_to_ls(pp_net.trafo["hv_bus"].values, pp_to_ls),
+                     pp_bus_to_ls(pp_net.trafo["lv_bus"].values, pp_to_ls))
 
     for tr_id, is_connected in enumerate(pp_net.trafo["in_service"].values):
         if not is_connected: