+++ src/GridModel.cpp
@@ -14,6 +14,8 @@ GridModel::GridModel(const GridModel & other)
 {
     reset(true, true, true);
 
+    _ls_to_pp = other._ls_to_pp;
+
     init_vm_pu_ = other.init_vm_pu_;
     sn_mva_ = other.sn_mva_;
 
@@ -78,6 +80,7 @@ GridModel::GridModel(const GridModel & other)
 GridModel::StateRes GridModel::get_state() const
 {
     std::vector<real_type> bus_vn_kv(bus_vn_kv_.begin(), bus_vn_kv_.end());
+    std::vector<int> ls_to_pp(_ls_to_pp.begin(), _ls_to_pp.end());
     int version_major = VERSION_MAJOR;
     int version_medium = VERSION_MEDIUM;
     int version_minor = VERSION_MINOR;
@@ -93,6 +96,7 @@ GridModel::StateRes GridModel::get_state() const
     GridModel::StateRes res(version_major,
                             version_medium,
                             version_minor,
+                            ls_to_pp,
                             init_vm_pu_,
                             sn_mva_,
                             bus_vn_kv,
@@ -132,30 +136,32 @@ void GridModel::set_state(GridModel::StateRes & my_state)
         exc_ << "It is not possible. Please reinstall it.";
         throw std::runtime_error(exc_.str());
     }
-    init_vm_pu_ = std::get<3>(my_state);
-    sn_mva_ = std::get<4>(my_state);
-    std::vector<real_type> & bus_vn_kv = std::get<5>(my_state);
-    std::vector<bool> & bus_status = std::get<6>(my_state);
+    std::vector<int> ls_to_pp_ = std::get<3>(my_state);
+    init_vm_pu_ = std::get<4>(my_state);
+    sn_mva_ = std::get<5>(my_state);
+    std::vector<real_type> & bus_vn_kv = std::get<6>(my_state);
+    std::vector<bool> & bus_status = std::get<7>(my_state);
 
     // powerlines
-    DataLine::StateRes & state_lines = std::get<7>(my_state);
+    DataLine::StateRes & state_lines = std::get<8>(my_state);
     // shunts
-    DataShunt::StateRes & state_shunts = std::get<8>(my_state);
+    DataShunt::StateRes & state_shunts = std::get<9>(my_state);
     // trafos
-    DataTrafo::StateRes & state_trafos = std::get<9>(my_state);
+    DataTrafo::StateRes & state_trafos = std::get<10>(my_state);
     // generators
-    DataGen::StateRes & state_gens = std::get<10>(my_state);
+    DataGen::StateRes & state_gens = std::get<11>(my_state);
     // loads
-    DataLoad::StateRes & state_loads = std::get<11>(my_state);
+    DataLoad::StateRes & state_loads = std::get<12>(my_state);
     // static gen
-    DataSGen::StateRes & state_sgens= std::get<12>(my_state);
+    DataSGen::StateRes & state_sgens= std::get<13>(my_state);
     // storage units
-    DataLoad::StateRes & state_storages = std::get<13>(my_state);
+    DataLoad::StateRes & state_storages = std::get<14>(my_state);
     // dc lines
-    DataDCLine::StateRes & state_dc_lines = std::get<14>(my_state);
+    DataDCLine::StateRes & state_dc_lines = std::get<15>(my_state);
 
     // assign it to this instance
 
+    _ls_to_pp =IntVect::Map(&ls_to_pp_[0], ls_to_pp_.size());
     // buses
     // 1. bus_vn_kv_
     bus_vn_kv_ = RealVect::Map(&bus_vn_kv[0], bus_vn_kv.size());